+++ flake.nix
@@ -14,7 +14,7 @@
       system:
       let
         name = "typatro";
-        version = "1.0.8";
+        version = "1.1.0";
 
         pkgs = import nixpkgs { inherit system; };
         python3 = pkgs.python312Packages;